linters fix dependency cicle
