+++ src/components/createList.js
@@ -1,6 +1,22 @@
 import { createTask, clearAllCompleted } from './crudList';
+import { data } from './statusUpdate';
+import listElement from './listElement';
 
-export default function createList() {
+const clearDOMList = () => {
+  const todoListElement = document.querySelectorAll('.list-element');
+  todoListElement.forEach((elm) => { elm.remove(); });
+};
+
+const drawListElements = () => {
+  const todoList = document.getElementById('todo-list');
+  data.forEach((task) => {
+    const element = listElement(task);
+    todoList.appendChild(element);
+  });
+  const clearBtn = document.getElementById('clear-btn');
+  todoList.appendChild(clearBtn);
+};
+const createList = () => {
   const todoList = document.getElementById('todo-list');
 
   /** ********************** First element **************************************** */
@@ -19,7 +35,13 @@ export default function createList() {
   const txtSearch = document.createElement('input');
   txtSearch.placeholder = 'Add to your list...';
   txtSearch.classList.add('txt-list');
-  txtSearch.addEventListener('keyup', createTask);
+  txtSearch.addEventListener('keyup', (e) => {
+    if (e.keyCode === 13) {
+      createTask(e);
+      clearDOMList();
+      drawListElements();
+    }
+  });
   searchElement.appendChild(txtSearch);
 
   /** ********************** End list element **************************************** */
@@ -28,11 +50,16 @@ export default function createList() {
   const btn = document.createElement('button');
   btn.innerText = 'Clear all elements';
   btn.classList.add('btn-clear');
-  btn.addEventListener('click', clearAllCompleted);
+  btn.addEventListener('click', () => {
+    clearAllCompleted();
+    clearDOMList();
+    drawListElements();
+  });
   cAllElement.appendChild(btn);
 
   todoList.appendChild(firstElement);
   todoList.appendChild(searchElement);
-
   todoList.appendChild(cAllElement);
-}
+};
+
+export { createList, drawListElements };
