+++ src/components/crudList.js
@@ -1,37 +1,16 @@
-/* eslint-disable import/no-cycle */
 import {
   data, addData, saveData, removeCompleted,
 } from './statusUpdate';
-import listElement from './listElement';
-import { orderData, clearData } from './interaction';
-
-const clearDOMList = () => {
-  const todoListElement = document.querySelectorAll('.list-element');
-  todoListElement.forEach((elm) => { elm.remove(); });
-};
-
-const runList = () => {
-  const todoList = document.getElementById('todo-list');
-  data.forEach((task) => {
-    const element = listElement(task);
-    todoList.appendChild(element);
-  });
-  const clearBtn = document.getElementById('clear-btn');
-  todoList.appendChild(clearBtn);
-};
+import { clearData } from './interaction';
 
 const createTask = (e) => {
-  if (e.keyCode === 13) {
-    const newTask = {
-      description: e.target.value,
-      completed: false,
-      index: data.length + 1,
-    };
-    addData(newTask);
-    clearDOMList();
-    runList();
-    saveData();
-  }
+  const lastItem = data[data.length - 1];
+  const newTask = {
+    description: e.target.value,
+    completed: false,
+    index: lastItem ? lastItem.index + 1 : 0,
+  };
+  addData(newTask);
 };
 
 const updateTask = (index, elm) => {
@@ -44,23 +23,16 @@ const clearAllCompleted = () => {
   const completed = data.filter((obj) => !obj.completed);
   clearData();
   removeCompleted(completed);
-  orderData();
   saveData();
-  clearDOMList();
-  runList();
 };
 
 const deleteTask = (index) => {
   const removed = data.filter((obj) => obj.index !== index);
   clearData();
   removeCompleted(removed);
-  orderData();
   saveData();
-  clearDOMList();
-  runList();
 };
 
 export {
-  createTask, runList, updateTask, clearAllCompleted, deleteTask,
+  createTask, updateTask, clearAllCompleted, deleteTask,
 };
-/* eslint-enable import/no-cycle */
