+++ src/index.js
@@ -1,12 +1,11 @@
 import './css/style.css';
 import { loadData } from './components/statusUpdate';
-import createList from './components/createList';
-import { runList } from './components/crudList';
+import { createList, drawListElements } from './components/createList';
 
 const run = () => {
   loadData();
   createList();
-  runList();
+  drawListElements();
 };
 
 run();
